code restructure / node error fix
changed the code so it follows the stylecop style except for things being public and order of provate / public variables

There was an error in the MyNode class where hovering over nodes with a disabled script would still run. Changed so when the script is disabled it does not run those methods.