Worked on bullet hit detection
I am attempting to shift our hit detection over to using colliders, this code is not yet functional but I am pushing so that I might get some help from the rest of the group. Note that I made bullets faster and enemies super weak in this build for testing.