Create Flame Tower #31
This breaks literally EVERY OTHER TOWER IN THE GAME, but it's fine. It has to be fine. It's the flame tower, the flame tower HAS to be in the game. Even at the cost of all other towers we NEED the flame tower. May god have mercy on your soul if you think we could cut the FLAME TOWER.